+++ src/Main/grpc.cpp
@@ -1084,8 +1084,8 @@ void upload_file(const GRPCReq *request, GRPCResp *response, SeriesWork *series)
 	}
 	if(request->has_content_length_header())
 	{
-		size_t content_length = stoul(request->header("Content-Length"));
-		size_t max_body_size = apiUtil->get_upload_max_body_size();
+		uint64_t content_length = stoul(request->header("Content-Length"));
+		uint64_t max_body_size = apiUtil->get_upload_max_body_size();
 		if (content_length > max_body_size)
 		{
 			SLOG_DEBUG("File size is " + to_string(content_length) + " byte, allowed max size " + to_string(max_body_size) + " byte!");
